feat: add support for prependBaseUrlToResourcesPath configuration
In this commit, a new configuration option `dbadmin.prepend-base-url-to-resources-path` is added. This option is used when there is a need to set the static resources path to the same as the application path. This might be useful when the application is hosted behind a reverse proxy with a prefix path.